+++ src/game/scenes/BootScene.ts
@@ -1,4 +1,5 @@
 import Phaser from "phaser";
+import { trackSessionStart } from "../meta/analytics";
 
 export class BootScene extends Phaser.Scene {
   constructor() {
@@ -7,6 +8,7 @@ export class BootScene extends Phaser.Scene {
 
   create(): void {
     this.cameras.main.setBackgroundColor("#0b3d3a");
+    trackSessionStart();
     this.scene.start("home");
   }
 }